+++ Assets/NumericalValueUIManager.cs
@@ -7,11 +7,11 @@ public class NumericalValueUIManager : MonoBehaviour
 {
     [SerializeField, Header("재화 표시")] private TextMeshProUGUI _moneyText;
     [SerializeField, Header("인구수 표시")] private TextMeshProUGUI _peopleCntText;
-    public Dictionary<NumericalValueType, Action<int>> valueUIDic=new();
+    public Dictionary<NumericalValueType, Action<float>> valueUIDic=new();
 
     private void Awake()
     {
-        valueUIDic = new Dictionary<NumericalValueType, Action<int>>
+        valueUIDic = new Dictionary<NumericalValueType, Action<float>>
         {
             { NumericalValueType.Money, OnMoneyUI},
             { NumericalValueType.Happiness, OnHappinessUI},
@@ -25,27 +25,27 @@ public class NumericalValueUIManager : MonoBehaviour
         NumericalValueManager.Instance.ModifyNumericalValue(NumericalValueType.Money, ModifyType.Add, 0);
     }
 
-    public void HandleValueChanged(NumericalValueType type, int value)
+    public void HandleValueChanged(NumericalValueType type, float value)
     {
         valueUIDic[type]?.Invoke(value);
     }
 
-    private void OnMoneyUI(int value)
+    private void OnMoneyUI(float value)
     {
         _moneyText.text = $"{value}원";
     }
 
-    private void OnPeopleCntUI(int value)
+    private void OnPeopleCntUI(float value)
     {
         _peopleCntText.text = $"{value}원";
     }
 
-    private void OnWarmingUI(int value)
+    private void OnWarmingUI(float value)
     {
 
     }
 
-    private void OnHappinessUI(int value)
+    private void OnHappinessUI(float value)
     {
 
     }
